Another try at fixing disabled color for lang menu (BL-13405,BL-13376)
